changed drag to u8, changed temp var to real var name

+++ lib/bombadil-schema/src/schema.rs
@@ -139,7 +139,7 @@ pub enum BrowserAction {
     MouseDrag {
         from: Point,
         to: Point,
-        steps: u32,
+        steps: u8,
         delay_millis: u64,
     },
     SetViewport {
